handling with mappstruct

+++ src/main/java/pl/devfinder/business/dao/OfferDAO.java
@@ -11,5 +11,5 @@ public interface OfferDAO {
 
     List<Offer> findAll();
 
-    Integer getNumberOfAvailableOffers(Long employerId);
+    Long getNumberOfAvailableOffers(Long employerId);
 }
